// src/types/enum/scene.ts
export enum SceneEnterReasonEnum {
  NONE = 0,
  LOGIN = 1,
  DUNGEON_REPLAY = 11,
  DUNGEON_REVIVE_ON_WAYPOINT = 12,
  DUNGEON_ENTER = 13,
  DUNGEON_QUIT = 14,
  GM = 21,
  QUEST_ROLLBACK = 31,
  REVIVAL = 32,
  PERSONAL_SCENE = 41,
  TRANS_POINT = 42,
  CLIENT_TRANSMIT = 43,
  FORCE_DRAG_BACK = 44,
  TEAM_KICK = 51,
  TEAM_JOIN = 52,
  TEAM_BACK = 53,
  MUIP = 54,
  DUNGEON_INVITE_ACCEPT = 55,
  LUA = 56,
  ACTIVITY_LOAD_TERRAIN = 57,
  HOST_FROM_SINGLE_TO_MP = 58,
  MP_PLAY = 59,
  ANCHOR_POINT = 60,
  LUA_SKIP_UI = 61,
  RELOAD_TERRAIN = 62,
  DRAFT_TRANSFER = 63,
  ENTER_HOME = 64,
  EXIT_HOME = 65,
  CHANGE_HOME_MODULE = 66,
  GALLERY = 67,
  HOME_SCENE_JUMP = 68,
  HIDE_AND_SEEK = 69,
  SUMMER_TIME_MIST_ZONE = 70,
  FORCE_QUIT_SCENE = 71,
  HERO_COURSE = 72,
  TRANS_CLIMATE = 73,
  SUMMER_TIME_BOAT_RESTART = 74,
  TSURUMI_MIST_CLIMATE = 75,
  TSURUMI_REGIONAL_MIST_CLIMATE = 76
}

export enum SceneEnterTypeEnum {
  ENTER_NONE = 0,
  ENTER_SELF = 1,
  ENTER_GOTO = 2,
  ENTER_JUMP = 3,
  ENTER_OTHER = 4,
  ENTER_BACK = 5,
  ENTER_DUNGEON = 6,
  ENTER_DUNGEON_REPLAY = 7,
  ENTER_GOTO_BY_PORTAL = 8,
  ENTER_SELF_HOME = 9,
  ENTER_OTHER_HOME = 10,
  ENTER_GOTO_RECREATE = 11
}